--- mbbk/mbur/src/main/java/com/mbti/user/user/config/auth/LoginUserArgumentResolver.java
@@ -1,31 +1,0 @@
-package com.mbti.user.user.config.auth;
-
-import com.mbti.user.user.config.auth.LoginUser;
-import com.mbti.user.user.user.dto.SessionUser;
-import jakarta.servlet.http.HttpSession;
-import lombok.RequiredArgsConstructor;
-import org.springframework.core.MethodParameter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@RequiredArgsConstructor
-@Component
-public class LoginUserArgumentResolver implements HandlerMethodArgumentResolver {
-
-    private final HttpSession httpSession;
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        boolean isLoginUserAnnotation = parameter.getParameterAnnotation(LoginUser.class) != null;
-        boolean isUserClass = SessionUser.class.equals(parameter.getParameterType());
-        return isLoginUserAnnotation && isUserClass;
-    }
-
-    @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
-        return httpSession.getAttribute("user");
-    }
-}
