--- mbbk/mbur/src/main/java/com/mbti/user/user/exception/BusinessException.java
@@ -1,10 +1,0 @@
-package com.mbti.user.user.exception;
-
-public class BusinessException extends RuntimeException {
-    public BusinessException(String message) {
-        super(message);
-    }
-    public BusinessException(String message, Throwable e) {
-        super(message, e);
-    }
-}
