--- mbbk/mbur/src/main/java/com/mbti/user/user/user/dto/Role.java
@@ -1,17 +1,0 @@
-package com.mbti.user.user.user.dto;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-
-    GUEST("ROLE_GUEST", "손님"),
-    USER("ROLE_USER", "일반 사용자");
-
-    private final String key;
-    private final String title;
-
-}
