--- mbbk/mbur/src/main/java/com/mbti/user/user/user/dto/SessionUser.java
@@ -1,22 +1,0 @@
-package com.mbti.user.user.user.dto;
-
-import com.mbti.user.user.user.entity.User;
-import lombok.Getter;
-
-import java.io.Serializable;
-
-/**
- * 직렬화 기능을 가진 User클래스
- */
-@Getter
-public class SessionUser implements Serializable {
-    private String name;
-    private String email;
-    private String picture;
-
-    public SessionUser(User user){
-        this.name = user.getName();
-        this.email = user.getEmail();
-        this.picture = user.getPicture();
-    }
-}
