--- mbbk/mbur/src/main/java/com/mbti/user/user/user/mapper/UserMapper.java
@@ -1,5 +1,0 @@
-package com.mbti.user.user.user.mapper;
-
-public interface UserMapper { //extends StructMapper<User, UserDto> {
-
-}
