--- mbbk/mbur/src/main/java/com/mbti/user/user/user/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.mbti.user.user.user.repository;
-
-import com.mbti.user.user.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, String> {
-    Optional<User> findByEmail(String email);
-}
