+++ mbbk/user/user/settings.gradle
@@ -1,0 +1,7 @@
+pluginManagement {
+	repositories {
+		maven { url 'https://repo.spring.io/milestone' }
+		gradlePluginPortal()
+	}
+}
+rootProject.name = 'user'
