+++ mbbk/user/user/src/main/java/com/mbti/user/UserApplication.java
@@ -1,0 +1,13 @@
+package com.mbti.user;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserApplication.class, args);
+	}
+
+}
