+++ mbbk/user/user/src/main/java/com/mbti/user/config/auth/LoginUser.java
@@ -1,0 +1,11 @@
+package com.mbti.user.config.auth;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LoginUser {
+}
