+++ mbbk/user/user/src/main/java/com/mbti/user/config/security/SecurityConfig.java
@@ -1,0 +1,34 @@
+package com.mbti.user.config.security;
+
+import com.mbti.user.user.dto.Role;
+import com.mbti.user.user.service.CustomOAuth2UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@Configuration
+@RequiredArgsConstructor
+@EnableWebSecurity // Spring Security 설정 활성화
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+
+    private final CustomOAuth2UserService customOAuth2UserService;
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+                .csrf().disable()
+                .headers().frameOptions().disable() // h2-console 화면을 사용하기 위해 해당 옵션 disable
+                .and()
+                .authorizeRequests()// URL별 권한 권리
+                .antMatchers("/**","/css/**","/images/**","/js/**","/h2-console/**").permitAll()
+                .antMatchers("/api/v1/**").hasRole(Role.USER.name()) // /api/v1/** 은 USER권한만 접근 가능
+                //.anyRequest().authenticated() // anyRequest : 설정된 값들 이외 나머지 URL 나타냄, authenticated : 인증된 사용자
+                //.anyRequest().permitAll()
+                .and()
+                .logout()
+                .logoutSuccessUrl("/");
+    }
+
+}
