+++ mbbk/user/user/src/main/java/com/mbti/user/exception/BusinessException.java
@@ -1,0 +1,10 @@
+package com.mbti.user.exception;
+
+public class BusinessException extends RuntimeException {
+    public BusinessException(String message) {
+        super(message);
+    }
+    public BusinessException(String message, Throwable e) {
+        super(message, e);
+    }
+}
