+++ mbbk/user/user/src/main/java/com/mbti/user/user/mapper/UserMapper.java
@@ -1,0 +1,5 @@
+package com.mbti.user.user.mapper;
+
+public interface UserMapper { //extends StructMapper<User, UserDto> {
+
+}
