+++ mbbk/user/user/src/main/java/com/mbti/user/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.mbti.user.user.repository;
+
+import com.mbti.user.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, String> {
+    Optional<User> findByEmail(String email);
+}
