+++ mbbk/user/user/src/test/java/com/mbti/user/UserApplicationTests.java
@@ -1,0 +1,13 @@
+package com.mbti.user;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
